Fix: Address multiple issues with tox, coverage report, temp files
Coverage report temporary files were interacting with tests that scanned and
processed local files and on disk. We now cehck the pyproject.toml stanza for
the configuration option, and provide a warning.

Tox is no longer invoked with all environments by default. As this is designed
to be called from a matrix job, it was resulting in duplicated/overlapping
test runs. We now run the tox environment matched to the Python version when
no environments have been specified.

Coverage reports have been swapped to XML format, so they can be processed or
uploaded into other types of platforms/tooling.

This update also addresses the path prefix handling and fixes the case of the
input values. We not explicitly install coverage[toml] to better support
coverage configuration from this central location.

Signed-off-by: Matthew Watkins <matt@modeseven.org>